+++ src/Symfony/Component/Debug/Tests/Fixtures2/RequiredTwice.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Symfony\Component\Debug\Tests\Fixtures2;
+
+class RequiredTwice
+{
+}