+++ src/Symfony/Component/DependencyInjection/Dumper/PhpDumper.php
@@ -904,9 +904,6 @@ EOF;
         \$this->services =
         \$this->scopedServices =
         \$this->scopeStacks = array();
-
-        \$this->set('service_container', \$this);
-
 EOF;
 
         $code .= "\n";