+++ src/Symfony/Component/Translation/Loader/FileLoader.php
@@ -46,7 +46,10 @@ abstract class FileLoader extends ArrayLoader
         }
 
         $catalogue = parent::load($messages, $locale, $domain);
-        $catalogue->addResource(new FileResource($resource));
+
+        if (class_exists('Symfony\Component\Config\Resource\FileResource')) {
+            $catalogue->addResource(new FileResource($resource));
+        }
 
         return $catalogue;
     }